% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doe.R
\name{doe_dataset}
\alias{doe_dataset}
\title{Get full details for a DOE EDX dataset}
\usage{
doe_dataset(id)
}
\arguments{
\item{id}{Dataset ID or name slug (e.g., "ree-and-coal-open-geodatabase-collection")}
}
\value{
tibble: single row with dataset details
}
\description{
Get full details for a DOE EDX dataset
}
