% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doe.R
\name{doe_resources}
\alias{doe_resources}
\title{List resources (downloadable files) for a DOE EDX dataset}
\usage{
doe_resources(id)
}
\arguments{
\item{id}{Dataset ID or name slug}
}
\value{
tibble: id, name, description, format, url, size, last_modified
}
\description{
List resources (downloadable files) for a DOE EDX dataset
}
